Create a blueprint

Create a blueprint

blueprint(
  name,
  command,
  description = NULL,
  metadata = NULL,
  export_metadata = TRUE,
  metadata_file_type = c("csv"),
  metadata_directory = here::here("blueprints"),
  metadata_file_path = NULL,
  ...,
  class = character()
)

Arguments

name
The name of the blueprint

command
The code to build the target dataset

description
An optional description of the dataset to be used for codebook generation

metadata
The associated variable metadata for this dataset

export_metadata
Indicator to render new metadata file for new dataset

metadata_file_type
The kind of metadata file. Currently only CSV.

metadata_directory
Where the metadata file will be stored.

metadata_file_path
Overrides the metadata file path generated by metadata_directory, name, and metadata_file_type if not NULL.

...
Any other parameters and settings for the blueprint

class
A subclass of blueprint capability, for future work

Value

A blueprint object

Cleanup Tasks

blueprintr offers some post-check tasks that attempt to match datasets to the metadata as much as possible. There are two default tasks that run:

Reorders variables to match metadata order.

Drops variables marked with dropped == TRUE if the dropped variable exists in the metadata.

The remaining tasks have to be enabled by the user:

If labelled = TRUE in the blueprint() command, all columns will be converted to labelled() columns, provided that at least the description field is filled in. If the coding column is present in the metadata, then categorical levels as specified by a coding() will be added to the column as well. In case the description field is used for detailed column descriptions, the title field can be added to the metadata to act as short titles for the columns.