Macros for blueprint authoring

blueprintr uses code inspection to identify and trace dataset dependencies. These macro functions signal a dependency to blueprintr and evaluate to symbols to be analyzed in the drake plan.

.TARGET(bp_name)

.BLUEPRINT(bp_name)

.META(bp_name)

Arguments

bp_name
Character string of blueprint's name

Functions

.TARGET: Gets symbol of built and checked data

.BLUEPRINT: Gets symbol of blueprint reference in plan

.META: Gets symbol of metadata reference in plan

When to use

Generally speaking, the .BLUEPRINT and .META macros should be used for check functions, which frequently require context, e.g. in the form of configuration from the blueprint or coding expectations from the metadata. .TARGET is primarily used in blueprint commands, but there could be situations where a check depends on the content of another dataset.

It is important to note that the symbols generated by these macros are only understood in the context of a drake plan. The targets associated with the symbols are generated when blueprints are attached to a plan.

Examples

.TARGET("example_dataset")

#> example_dataset

.BLUEPRINT("example_dataset")

#> example_dataset_blueprint

.META("example_dataset")

#> example_dataset_meta

blueprint(
  "test_bp",
  description = "Blueprint with dependencies",
  command =
    .TARGET("parent1") %>%
      left_join(.TARGET("parent2"), by = "id") %>%
      filter(!is.na(id))
)

#> <blueprint: 'test_bp'>
#> 
#> Description: Blueprint with dependencies
#> Metadata location: '/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmpaut8r2/file6cd3f2119b9/blueprints/test_bp.csv'
#> 
#> -- Command --
#> Drake command:
#> parent1 %>% left_join(parent2, by = "id") %>% filter(!is.na(id))
#> 
#> Raw command:
#> .TARGET("parent1") %>% left_join(.TARGET("parent2"), by = "id") %>% 
#>     filter(!is.na(id))